Updated Test Controller and Implemented a
Got rid of the greetings so only /test is getting hit by Fetch API. Created an html file for the frontend that uses Bootstrap's CSS and JS bundles loaded from a CDN which gives us easy access to styling tools and interactive components like dropdowns. Implemented a JS alert to simulate a form submission and added a test button to call the /test endpoint.